# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GlobalHR Platform API - HR, Payroll, Contracts, Compliance, Onboarding
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workers:
    get:
      operationId: listWorkers
      tags: [workers]
      summary: List all workers
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, inactive, onboarding, terminated]
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [employee, contractor, eor]
        - name: country
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of workers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Worker"
    post:
      operationId: createWorker
      tags: [workers]
      summary: Create a worker
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkerBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Worker"
  /workers/{id}:
    get:
      operationId: getWorker
      tags: [workers]
      summary: Get a worker by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Worker
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Worker"
        "404":
          description: Not found
    patch:
      operationId: updateWorker
      tags: [workers]
      summary: Update a worker
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkerBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Worker"
        "404":
          description: Not found
    delete:
      operationId: deleteWorker
      tags: [workers]
      summary: Terminate a worker
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Terminated

  # ─── CONTRACTS ────────────────────────────────────────────────────────────
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payment runs
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, processing, completed, failed]
        - name: workerId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Create a payment run
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get a payment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "404":
          description: Not found
  /payments/{id}/approve:
    post:
      operationId: approvePayment
      tags: [payments]
      summary: Approve a payment run
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  # ─── COMPLIANCE ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Worker:
      type: object
      properties:
        id:
          type: integer
        organizationId:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        jobTitle:
          type: string
        department:
          type: string
        workerType:
          type: string
          enum: [employee, contractor, eor]
        status:
          type: string
          enum: [active, inactive, onboarding, terminated]
        country:
          type: string
        currency:
          type: string
        salary:
          type: ["number", "null"]
        avatarUrl:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
          format: date
        endDate:
          type: ["string", "null"]
          format: date
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, organizationId, firstName, lastName, email, jobTitle, department, workerType, status, country, currency, createdAt, updatedAt]
    CreateWorkerBody:
      type: object
      properties:
        organizationId:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        jobTitle:
          type: string
        department:
          type: string
        workerType:
          type: string
          enum: [employee, contractor, eor]
        country:
          type: string
        currency:
          type: string
        salary:
          type: ["number", "null"]
        avatarUrl:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
      required: [organizationId, firstName, lastName, email, jobTitle, department, workerType, country, currency]
    UpdateWorkerBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        jobTitle:
          type: string
        department:
          type: string
        status:
          type: string
          enum: [active, inactive, onboarding, terminated]
        salary:
          type: ["number", "null"]
        avatarUrl:
          type: ["string", "null"]

    # ─── CONTRACT ────────────────────────────────────────────────────────────
...
    Payment:
      type: object
      properties:
        id:
          type: integer
        workerId:
          type: integer
        organizationId:
          type: integer
        amount:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [pending, processing, completed, failed]
        paymentType:
          type: string
          enum: [salary, bonus, expense, invoice]
        periodStart:
          type: ["string", "null"]
          format: date
        periodEnd:
          type: ["string", "null"]
          format: date
        processedAt:
          type: ["string", "null"]
          format: date-time
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, workerId, organizationId, amount, currency, status, paymentType, createdAt, updatedAt]
    CreatePaymentBody:
      type: object
      properties:
        workerId:
          type: integer
        organizationId:
          type: integer
        amount:
          type: number
        currency:
          type: string
        paymentType:
          type: string
          enum: [salary, bonus, expense, invoice]
        periodStart:
          type: ["string", "null"]
        periodEnd:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required: [workerId, organizationId, amount, currency, paymentType]

    # ─── COMPLIANCE ───────────────────────────────────────────────────────────
...